create company service (and some refactoring)
