+++ src/services/job.ts
@@ -1,3 +1,5 @@
+import { addCompanyDataInJob } from "./company";
+
 export const getRecentJobs = async () => {
   try {
     const res = await fetch(
@@ -38,35 +40,5 @@ export const createJob = async (data) => {
   } catch (error) {
     console.log(error);
     return "Erro";
-  }
-};
-
-const addCompanyDataInJob = async (jobs) => {
-  const companiesId = jobs.map((job) => job.company);
-  const companiesData = await getCompaniesData(companiesId);
-  const jobsWithCompanyData = jobs.map((job) => ({
-    ...job,
-    companyData: companiesData.find((company) => company.id === job.company),
-  }));
-  return jobsWithCompanyData;
-};
-
-const getCompaniesData = async (companiesId: string[]) => {
-  const companiesQuery = companiesId.map(
-    (companyId: string) => `id=${companyId}&`
-  );
-  const companiesQueryUrl = `?${companiesQuery.join("").slice(0, -1)}`; //Slice remove the  last "&'
-  console.log(companiesQueryUrl);
-  try {
-    const res = await fetch(
-      `http://localhost:4000/companies${companiesQueryUrl}`
-    );
-
-    if (res.ok) {
-      const data = await res.json();
-      return data;
-    }
-  } catch (error) {
-    console.log(error);
   }
 };
